+++ src/AngleSharp/Html/Parser/HtmlDomBuilder.cs
@@ -3084,6 +3084,12 @@ namespace AngleSharp.Html.Parser
 
                 furthestBlock.AddNode(element);
                 _formattingElements.Remove(formattingElement);
+
+                if (bookmark > _formattingElements.Count)
+                {
+                    bookmark = _formattingElements.Count;
+                }
+
                 _formattingElements.Insert(bookmark, element);
                 CloseNode(formattingElement);
                 _openElements.Insert(_openElements.IndexOf(furthestBlock) + 1, element);